New core objects and data model Remove sample integration test

+++ test/server/utils/hash_test.cc
@@ -22,7 +22,7 @@ TEST(hash_test, sha3_256_hash_test) {
   std::string my_hash(32, '\0');
   utils::sha3_256((unsigned char *) &my_hash.at(0), nullptr, 0);
   std::string res_hash(64, '\0');
-  for (int i = 0; i < my_hash.size(); ++i){
+  for (size_t i = 0; i < my_hash.size(); ++i){
     sprintf(&res_hash[2 * i], "%02x", my_hash[i] & 0xFF);
   }
   ASSERT_STREQ(res_hash.c_str(), "a7ffc6f8bf1ed76651c14756a061d662f580ff4de43b49fa82d80a4b80f8434a");
@@ -31,7 +31,7 @@ TEST(hash_test, sha3_256_hash_test) {
   std::string hash(32, '\0');
 
   utils::sha3_256((unsigned char *) &hash.at(0), blob.data(), 32);
-  for (int i = 0; i < hash.size(); ++i){
+  for (size_t i = 0; i < hash.size(); ++i){
     printf("%02x", hash[i] & 0xFF);
   }
 }
